WIP: Fix NuSTAR example gallery (#199)
* (1) high_break needs to be cast to float64 (or something higher) as numpy does not do implicitly anymore and it causes nans; (2) index->ind again for the warm target model to keep it unique from the cold thick target inputs; and (3) separate working NuSTAR examples in the example gallery.

---------
